Yandex cup 2020 Resolver is done
